+++ tests/test_matching.py
@@ -2,13 +2,15 @@ import pytest
 
 import missive as m
 
+from .matchers import always, never
+
 
 def test_one_matching_handler():
     processor: m.Processor[m.GenericMessage] = m.Processor()
 
     flag = False
 
-    @processor.handle_for([])
+    @processor.handle_for(always)
     def flip_bit(
         message: m.GenericMessage, ctx: m.HandlingContext[m.GenericMessage]
     ) -> None:
@@ -29,7 +31,7 @@ def test_one_matching_handler():
 def test_no_matching_handler():
     processor: m.Processor[m.GenericMessage] = m.Processor()
 
-    @processor.handle_for((lambda m: False,))
+    @processor.handle_for(never)
     def non_matching_handler(message, ctx):
         assert False
 
@@ -46,11 +48,11 @@ def test_no_matching_handler():
 def test_multiple_matching_handlers():
     processor: m.Processor[m.GenericMessage] = m.Processor()
 
-    @processor.handle_for((lambda m: True,))
+    @processor.handle_for(always)
     def a_matching_handler(message, ctx):
         message.ack()
 
-    @processor.handle_for((lambda m: True,))
+    @processor.handle_for(always)
     def another_matching_handler(message, ctx):
         message.ack()
 
@@ -69,14 +71,14 @@ def test_one_matching_handler_among_multiple():
 
     flag = False
 
-    @processor.handle_for((lambda m: True,))
+    @processor.handle_for(always)
     def a_matching_handler(message, ctx):
         nonlocal flag
         flag = True
         ctx.ack(message)
 
-    @processor.handle_for((lambda m: False,))
-    def another_matching_handler(message, ctx):
+    @processor.handle_for(never)
+    def another_handler(message, ctx):
         message.ack()
 
     test_client = processor.test_client()